Store Compiled Packet Parsers in a global cache - Resolves #722
When a connection is closed, any compiled parser created for that will
be lost, as the connection is removed.

However, Node/V8 Internals appear to be holding onto any created function.

This can result in memory leaking based on connection cycling and new
parsers being created again.

This change creates a global cache store since everything around the generated
code is not connection specific.

This also completes a TODO to use an LRU cache, although I set a high value
as to avoid running into the leak if the LRU is too small.

An API is provided for anyone who does need to control the max value, as
well as an API to clear the parser cache.